requirements.txt
discord.py
python-dotenv
text-to-owo
selenium
grpcio
Pillow